% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmt-package.R
\docType{package}
\name{dlmt-package}
\alias{dlmt-package}
\alias{dlmt}
\title{The 'dlmt' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2021). RStan: the R interface to Stan. R package version 2.21.3. https://mc-stan.org
}
