% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmt.R
\name{dlmt}
\alias{dlmt}
\title{Run dynamic linear model with transformations (Che and Glickman, 2023+)}
\usage{
dlmt(
  df,
  outcome = y,
  time = t,
  event = g,
  unit = a,
  centering = c("median", "mean", "none"),
  optimize = T,
  optimize_method = c("optim", "mcmc"),
  lam = NULL,
  model_params = NULL,
  smooth = F,
  details = F,
  ...
)
}
\arguments{
\item{df}{dataframe}

\item{outcome}{(unquoted) name of outcome column}

\item{time}{(unquoted) name of time column}

\item{event}{(unquoted) name of event ID column}

\item{unit}{(unquoted) name of unit ID column}

\item{centering}{centering to apply to outcomes from each event: "median", "mean", or "none"}

\item{optimize}{find an optimal outcome transformation and innovation variance parameter?}

\item{optimize_method}{method for estimating optimal parameters: "optim" for MAP estimation, "mcmc" for full MCMC estimation}

\item{lam}{(optional) vector of transformation parameters lambda; default to identity transformation. Note that the given lambda parameters will be used on the I-spline basis associated with the training data in the given df}

\item{model_params}{vector of model parameters (w, V0, a0, b0): (innovation variance, initial variance, shape and scale parameters for inverse-gamma prior on observation variance)}

\item{smooth}{smooth posterior latent ability estimates?}

\item{details}{include visualization of transformation, model predictions, and model estimates in output?}

\item{...}{optional parameters, e.g., number of knots, number of training periods}
}
\value{
df with posterior model estimates at each time period, optionally with helpful attributes
}
\description{
Run dynamic linear model with transformations (Che and Glickman, 2023+)
}
