% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_functions.R
\name{get_ispline_basis}
\alias{get_ispline_basis}
\title{Generate monotone (I-spline) spline basis}
\usage{
get_ispline_basis(y, quantiles = c(0.25, 0.5, 0.75), degree = 3, knots = NULL)
}
\arguments{
\item{y}{outcome vector}

\item{quantiles}{quantiles of y to use as spline basis knots}

\item{degree}{spline degree}

\item{knots}{optional; manually specified spline basis knots}
}
\value{
I-spline basis matrix
}
\description{
Generate monotone (I-spline) spline basis
}
