% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_functions.R
\name{get_ispline_coefs}
\alias{get_ispline_coefs}
\title{Get I-spline coefficients from y ~ basis spline fit}
\usage{
get_ispline_coefs(y, basis, nonneg = T)
}
\arguments{
\item{y}{outcome vector}

\item{basis}{I-spline basis}

\item{nonneg}{restrict coefficients to be non-negative?}
}
\value{
monotone spline coefficient vector (intercept, coefs)
}
\description{
Get I-spline coefficients from y ~ basis spline fit
}
