% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Generate useful keys and optionally center data}
\usage{
preprocess_data(
  df,
  outcome = y,
  time = t,
  event = g,
  unit = a,
  event_center = c("none", "median", "mean"),
  model_class = c("multi", "h2h")
)
}
\arguments{
\item{df}{raw data: columns unitID, eventID, t, y}

\item{outcome}{outcome variable}

\item{time}{time variable}

\item{event}{event variable}

\item{unit}{unit variable}

\item{event_center}{name of event-centering method: "none", "median", or "mean"}

\item{model_class}{class of model, either "multi" for multiplayer model or "h2h" for head-to-head model}
}
\value{
preprocessed df with helpful attributes
}
\description{
Generate useful keys and optionally center data
}
