% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmt_helpers.R
\name{run_dlmt}
\alias{run_dlmt}
\title{Run no-intercept DLM}
\usage{
run_dlmt(
  df,
  outcome = y,
  time = t,
  event = g,
  unit = a,
  model_params,
  model_class = c("multi", "h2h"),
  clear_diags = F,
  cap_var = T
)
}
\arguments{
\item{df}{data to run model on}

\item{outcome}{(unquoted) outcome variable}

\item{time}{(unquoted) time variable}

\item{event}{(unquoted) event variable}

\item{unit}{(unquoted) unit variable}

\item{model_params}{model hyperparameters w, v0, a0, b0}

\item{model_class}{class of model, either multiplayer ("multi") or head-to-head ("h2h")}

\item{clear_diags}{clear diagonals of V matrix between time periods?}

\item{cap_var}{cap maximum variance for players at initial variance?}
}
\value{
many lists of model results
}
\description{
Run no-intercept DLM
}
