% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Given true parameters, generate synthetic data according to DLM model}
\usage{
sim_data(
  W = 0.5,
  V0 = 10,
  sd = 5,
  inv_trans_fun = function(x) {
     x
 },
  num_athletes = 10,
  ath_per_game = 5,
  num_games = 10,
  tmax = 6,
  HX = T
)
}
\arguments{
\item{W}{ratio of innovation variance to observation variance}

\item{V0}{ratio of intial player ability variance to observation variance}

\item{sd}{observation standard deviation}

\item{inv_trans_fun}{(inverse of) transformation to apply to data}

\item{num_athletes}{total number of athletes}

\item{ath_per_game}{number of athletes that play each game}

\item{num_games}{number of games per time period}

\item{tmax}{number of time periods}

\item{HX}{generate scores around centered thetas?}
}
\value{
A synthetic dataset with outcomes y, time periods t, games (events) g, and athletes (units) a
}
\description{
Given true parameters, generate synthetic data according to DLM model
}
