% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlmt.R
\name{dlmt}
\alias{dlmt}
\title{Run dynamic linear model with transformations (Che and Glickman, 2023+)}
\usage{
dlmt(
  df,
  outcome = y,
  time = t,
  event = g,
  unit = a,
  centering = c("median", "mean", "none"),
  optimize = T,
  optimize_method = c("optim", "mcmc"),
  lam = NULL,
  model_params = NULL,
  smooth = F,
  details = F,
  ...
)
}
\arguments{
\item{df}{dataframe}

\item{outcome}{(unquoted) name of outcome column}

\item{time}{(unquoted) name of time column}

\item{event}{(unquoted) name of event ID column}

\item{unit}{(unquoted) name of unit ID column}

\item{centering}{centering to apply to outcomes from each event: "median", "mean", or "none"}

\item{optimize}{find an optimal outcome transformation and innovation variance parameter?}

\item{optimize_method}{method for estimating optimal parameters: "optim" for MAP estimation, "mcmc" for full MCMC estimation}

\item{lam}{(optional) vector of transformation parameters lambda; default to identity transformation. Note that the given lambda parameters will be used on the I-spline basis associated with the training data in the given df}

\item{model_params}{vector of model parameters (w, V0, a0, b0): (innovation variance, initial variance, shape and scale parameters for inverse-gamma prior on observation variance)}

\item{smooth}{smooth posterior latent ability estimates?}

\item{details}{include visualization of transformation, model predictions, and model estimates in output?}

\item{...}{optional parameters, e.g., number of knots, number of training periods}
}
\value{
df with posterior model estimates at each time period:
\itemize{
\item \code{time}: (user-defined name) time period of posterior estimates
\item \code{a_posts}: shape parameter of posterior inverse-gamma distribution of observation variance
\item \code{b_posts}: scale parameter of posterior inverse-gamma distribution of observation variance
\item \code{m_posts}: mean of posterior normal distribution of athlete latent abilities,
for athletes who participate in any games in the time period
\item \code{V_posts}: covariance matrix of posterior normal distribution of athlete latent abilities,
for athletes who participate in any games in the time period
\item \code{X_mats}: X matrix used for computations for given time period
\item \code{ptcps}: vector of shortened athlete IDs indicating athletes who participate in the given time period
}

Attributes of output df:
\itemize{
\item \code{model_params}: vector (w, V0, a0, b0) of parameters used to fit the model
\item \code{lambda}: vector of transformation parameters used for monotone spline transformation
}

Optional attributes (details == T):
\itemize{
\item \code{transformation_plot}: ggplot visualizing transformation
\item \code{predictions}: data frame of observed and predicted values for each unit-event outcome.
\itemize{
\item pt_est: point estimate of outcome
\item sd: standard deviation of point estimate
\item sd_theta: standard deviation of latent ability estimate (ignoring observation variance)
}
\item \code{m_values}: posterior means of latent ability estimates for each unit
\item \code{V_values}: posterior variance matrices of unit latent abilities relative to observation variance
\item \code{sig_values}: shape/scale of posterior distribution of observation variance at final time point,
along with its mean and mode.
}
}
\description{
Run dynamic linear model with transformations (Che and Glickman, 2023+)
}
